rename components to follow AirBNB guidelines
https://github.com/airbnb/javascript/tree/master/react#naming
