+++ website/src/components/Header.js
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import styles from "./header.module.scss";
+import styles from "./Header.module.scss";
 
 export default function Footer() {
   return (
