--- website/src/components/post/post.module.scss
@@ -1,39 +1,0 @@
-@import "styles/_scaffold.scss";
-
-.container {
-}
-
-.post-header {
-  h1 {
-    color: $primary;
-    font-size: 30px;
-
-    @include media-breakpoint-up(md) {
-      font-size: 60px;
-    }
-
-    a {
-      text-decoration: none;
-    }
-  }
-}
-
-.divider {
-  padding-right: 5px;
-  padding-left: 5px;
-}
-
-.post-footer {
-  margin-top: 20px;
-}
-
-.meta {
-  margin-bottom: 0;
-  font-family: $font-sans;
-  font-size: 14px;
-  text-transform: uppercase;
-
-  @include media-breakpoint-up(md) {
-    font-size: 16px;
-  }
-}
